backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
groq>=0.11.0
edge-tts>=6.1.9
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
aiofiles>=24.1.0

